+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - enhance: updated many NPC names barter on Find Tab
 = enhance: updated translations of skills and Find Tab data for RU up to U47.1
 - fix: fixed travel to kinship member's house in FR after unannounced renaming of skill by SSG
+- fix: plugin now loads again for creep characters when travel skills have no acquisition data
 
 ## v4.6.0